requirements.txt
eel==0.16.0
openpyxl==3.1.2
pywin32==306
pandas==2.2.0
requests==2.31.0
pillow==10.2.0
pyinstaller==6.4.0
